+++ Editor/Settings/AssetLintingSettingsEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a84ad664916343a082846503a82d0b8c
+timeCreated: 1629154637
